+++ NEWS.md
@@ -3,6 +3,22 @@ This file documents the development of the package as well as open issues or poi
 
 
 
+
+### Version in 3.6.0
+Date: 15.05.2026 - in progress
+
+This release includes autoselect methods for shash, betabinomial and taylor
+models using parallel code execution.
+
+Changes:
+
+*    new methods: autoselect.betabinomial, autoselect.shash
+
+
+
+
+
+
 ### Version in 3.5.4
 Date: 15.05.2026 - release